refactored: update-password functionality separated from the reset password page and refactored their codes.
The reset password page will be the page where the user will provide their email to.

Once to reset password is clicked, it will send the email to the server and the supabase server will send the recovery link to the user.

After the recovery link is clicked from the email, it will redirect to the update-password page where the user will update their password.