Fix realtime on conversation list, Fix unread message count,  Change message bubble UI.
Fix an issue on conversations list wherein reload is needed for the new message to reflect.

Fix an issue wherein the unread message count doesn't reflect proper value and doesn't dissappear when the message is read unless the receiver sends a new message.

Change the color them of the message to match the Visconn's theme.