Add email verification overlay
Add email verification overlay telling the user's to check their email for both of the singup and the forgot-password attempts.

Modify the return statements of both the register and forgot-password server functions as it automatically refreshes the page. The solution is to make an api route for both functions to handle the return statements.